+++ projects/journey-maps-client-testapp/karma.conf.js
@@ -27,6 +27,14 @@ module.exports = function (config) {
     colors: true,
     logLevel: config.LOG_INFO,
     autoWatch: true,
+    // https://flow.sbb.ch/browse/CLEW-13224
+    browsers: ['ChromeHeadlessNoSandbox'],
+    customLaunchers: {
+      ChromeHeadlessNoSandbox: {
+        base: 'ChromeHeadless',
+        flags: ['--no-sandbox', '--disable-setuid-sandbox']
+      }
+    },
     singleRun: false,
     restartOnFileChange: true
   });